requirements.txt
email-validator==2.2.0
fastapi==0.115.0
pydantic==2.9.2
pydantic-core==2.23.4
numpy>=1.26,<2.0
uvicorn==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.5.2
redis==5.0.8
httpx==0.27.2
alembic==1.13.2
python-multipart==0.0.9
pytest==8.3.3
flake8==7.1.1
mypy==1.11.2
slowapi==0.1.9
sentencepiece==0.2.1
